+++ lib/streamaggr/dedup_timing_test.go
@@ -65,7 +65,7 @@ func newBenchSamples(count int) []pushSample {
 			Name:  "app",
 			Value: fmt.Sprintf("instance-%d", i),
 		})
-		keyBuf = compressLabels(keyBuf[:0], labels[:labelsLen], labels[labelsLen:])
+		keyBuf = compressLabels(keyBuf[:0], labels[:labelsLen], labels[labelsLen:], true)
 		sample.key = string(keyBuf)
 		sample.value = float64(i)
 	}